- prepareConfig now returns an array of errors - fiddle with tests on parsing config files - remove statics from arg parsing - change Delegate().Address() to Address()

+++ briefcase/briefcase_test.go
@@ -154,7 +154,7 @@ func TestResetBriefcase(t *testing.T) {
 		"b": true,
 	}
 
-	sversioned := map[string]int{
+	sversioned := map[string]int64{
 		"a": 3,
 		"b": 8,
 	}
